Fetch and parse profiles from api
Breaking down the task into parts. Networking layer is implemented and tested to
fetch and parse the json payload. We are reading flyweight profile details as
we dont have hard requirements for the UI as yet.

Next tasks would be to build the UI to expose the Profiles and build the Room
database for querying and updating the status of user's choice.

Day 1