Store data in db
This builds the Data layers (repo, dao, db). Following the MVVM pattern data is
accesible to the VM only with all the business logic. Thus testing the VM by
mocking dependencies is possible.

Next stages include building the UI layers to the VM and testing the application.
Will be exploring edge cases, if any, and testing to ensure they are handled as
expected.

Day 1, end