UI implementation for the card views
Card ui with two buttons to accept/decline profiles. Buttons are replaced by
text stating users choice and the state is persisted in the DB.

This concludes all the requirements of the assignment.

There are various notes as comments in the code stating why some code decisions
were made and alternatives that can be used in case of changed requirements.
Further work would include working closely with the designer to ensure UI looks
acceptable and make any tweaks as suggested.

- Regular buttons were used rather than Image buttons for now.
- Didn't use CardView library for now. Can be added if necessary.

Day 2