Fix changing matchStatus of terminal entries in list doesnt update the VH's height
On accepting or declining a match of a profile in the second last position, the
view holder's height change is not propagated to the RV if the list is scrolled
all the way to the end.

Fallback to rebinding UI on match status change using RV.adapter's notify item
changed method. This causes the onMeasure to be called and RV to recalculate the
VH's height.